# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing shared expenses, groups, and user authentication",
        "title": "Qashare API",
        "contact": {
            "name": "Pranaov S",
            "email": "qashare.contact@pranaovs.me"
        },
        "license": {
            "name": "AGPL-3.0",
            "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the authenticated user's profile information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "me"
                ],
                "summary": "Get current user",
                "responses": {
                    "200": {
                        "description": "Returns the authenticated user's profile information",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "401": {
                        "description": "INVALID_TOKEN: Authentication token is missing, invalid, or expired",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "404": {
                        "description": "USER_NOT_FOUND: The authenticated user no longer exists in the database",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal server error - unexpected database error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the authenticated user's editable details. This is a full replacement, so all required fields (name and email) must be provided. Immutable fields will be ignored if included in the request body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "me"
                ],
                "summary": "Update current user (full replacement)",
                "parameters": [
                    {
                        "description": "Updated user details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns updated user",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST: Invalid request body or missing required fields",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "401": {
                        "description": "INVALID_TOKEN: Authentication token is missing, invalid, or expired",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "404": {
                        "description": "USER_NOT_FOUND: The authenticated user no longer exists",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal server error - unexpected database error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update specific fields of the authenticated user. Only provided fields are updated, others remain unchanged. Immutable fields (like user_id) will be ignored if included in the request body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "me"
                ],
                "summary": "Partially update current user",
                "parameters": [
                    {
                        "description": "Partial user details (name and/or email, all optional)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns updated user",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST: Invalid request body or validation failed",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "401": {
                        "description": "INVALID_TOKEN: Authentication token is missing, invalid, or expired",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "404": {
                        "description": "USER_NOT_FOUND: The authenticated user no longer exists",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal server error - unexpected database error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    }
                }
            }
        },
        "/v1/me/admin": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all groups that the authenticated user created (is admin of)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "me"
                ],
                "summary": "List groups user administers",
                "responses": {
                    "200": {
                        "description": "Returns list of groups the user is admin of",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Group"
                            }
                        }
                    },
                    "401": {
                        "description": "INVALID_TOKEN: Authentication token is missing, invalid, or expired",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal server error - unexpected database error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    }
                }
            }
        },
        "/v1/me/groups": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all groups the logged in user is a member of",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "me"
                ],
                "summary": "List user's groups",
                "responses": {
                    "200": {
                        "description": "Returns list of groups the user is a member of",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Group"
                            }
                        }
                    },
                    "401": {
                        "description": "INVALID_TOKEN: Authentication token is missing, invalid, or expired",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal server error - unexpected database error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apierrors.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "e.g., \"BAD_NAME\", \"INVALID_EMAIL\"",
                    "type": "string"
                },
                "message": {
                    "description": "Human-readable message",
                    "type": "string"
                }
            }
        },
...
        "models.Group": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "created_by": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "guest": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}